+++ app/Http/Controllers/RatingController.php
@@ -10,13 +10,11 @@ use OpenApi\Attributes as OAT;
 use Kami\Cocktail\OpenAPI as BAO;
 use Kami\Cocktail\Models\Cocktail;
 use Kami\Cocktail\Http\Requests\RatingRequest;
-use Illuminate\Http\Resources\Json\JsonResource;
-use Kami\Cocktail\Http\Resources\RatingResource;
 
 class RatingController extends Controller
 {
-    #[OAT\Post(path: '/ratings/cocktails/{id}', tags: ['Ratings'], summary: 'Rate a cocktail', parameters: [
-        new BAO\Parameters\BarIdParameter(),
+    #[OAT\Post(path: '/cocktails/{id}/ratings', tags: ['Ratings'], summary: 'Rate a cocktail', parameters: [
+        new BAO\Parameters\DatabaseIdParameter(),
     ], requestBody: new OAT\RequestBody(
         required: true,
         content: [
@@ -25,37 +23,36 @@ class RatingController extends Controller
             ]),
         ]
     ))]
-    #[OAT\Response(response: 200, description: 'Successful response', content: [
-        new BAO\WrapObjectWithData(BAO\Schemas\Rating::class),
-    ])]
+    #[OAT\Response(response: 201, description: 'Successful response')]
+    #[BAO\NotFoundResponse]
     #[BAO\NotAuthorizedResponse]
-    public function rateCocktail(RatingRequest $request, int $cocktailId): JsonResource
+    public function rateCocktail(RatingRequest $request, int $id): Response
     {
-        $cocktail = Cocktail::findOrFail($cocktailId);
+        $cocktail = Cocktail::findOrFail($id);
 
         if ($request->user()->cannot('rate', $cocktail)) {
             abort(403);
         }
 
-        $rating = $cocktail->rate(
+        $cocktail->rate(
             (int) $request->post('rating'),
             $request->user()->id
         );
 
         $cocktail->searchable();
 
-        return new RatingResource($rating);
+        return new Response(null, 201);
     }
 
-    #[OAT\Delete(path: '/ratings/cocktails/{id}', tags: ['Ratings'], summary: 'Delete current user\'s cocktail rating', parameters: [
+    #[OAT\Delete(path: '/cocktails/{id}/ratings', tags: ['Ratings'], summary: 'Delete current user\'s cocktail rating', parameters: [
         new BAO\Parameters\DatabaseIdParameter(),
     ])]
     #[OAT\Response(response: 204, description: 'Successful response')]
     #[BAO\NotAuthorizedResponse]
     #[BAO\NotFoundResponse]
-    public function deleteCocktailRating(Request $request, int $cocktailId): Response
+    public function deleteCocktailRating(Request $request, int $id): Response
     {
-        $cocktail = Cocktail::findOrFail($cocktailId);
+        $cocktail = Cocktail::findOrFail($id);
 
         if ($request->user()->cannot('rate', $cocktail)) {
             abort(403);